Move SupportsReplaceNumClasses to module_interfaces.py
