+++ src/super_gradients/modules/__init__.py
@@ -29,7 +29,7 @@ from super_gradients.modules.detection_modules import (
     SSDBottleneckNeck,
     SSDHead,
 )
-from super_gradients.modules.interfaces import SupportsReplaceNumClasses
+from super_gradients.module_interfaces import SupportsReplaceNumClasses
 
 __all__ = [
     "BaseDetectionModule",
